feat(cron): auto-refresh gold price at 9am and fund NAVs at 6pm VN time
- Extract scrapeGoldPrice() and scrapeFundNav() into shared lib modules
- Create /api/cron/refresh-gold (runs 02:00 UTC = 9am VN, updates all users)
- Create /api/cron/refresh-navs (runs 11:00 UTC = 6pm VN, updates all funds)
- Add vercel.json with cron schedules
- Both cron routes use service role client (bypasses RLS) and CRON_SECRET auth
